backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "helma-reveal-backend"
version = "0.1.0"
description = "Python backend for Helma stencil/glitter reveal processing"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "python-multipart>=0.0.9",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.6,<3",
    "Pillow>=10.4,<11",
    "opencv-python-headless>=4.10,<5",
    "numpy>=1.26,<2",
    "scikit-image>=0.24,<1",
    "rembg>=2.0.57,<3",
    "svgwrite>=1.4,<2",
    "reportlab>=4.2,<5",
    "qrcode[pil]>=7.4,<8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
]
face = [
    "mediapipe>=0.10,<1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
